add user based showing of products

+++ app/api/products/[id]/route.js
@@ -1,17 +1,37 @@
 import { NextResponse } from "next/server";
 import connectToDB from "../../database/connectToDB";
 import Product from "../../models/Product";
+import { findIfUserAuthenticated } from "../../customMiddleware";
 
 export async function PATCH(request, { params }) {
   try {
     await connectToDB();
     const body = await request.json();
     const { id } = params;
+
+    const { data, authenticateResponse } = findIfUserAuthenticated(request);
+    // this will be returned if it is set in the middleware
+    if (authenticateResponse != null) {
+      return authenticateResponse;
+    }
+
     // find the product with that id
     const product = await Product.findById(id);
     if (!product) {
       return NextResponse.json(
         { success: false, message: "Error while updating the product" },
+        { status: 404 }
+      );
+    }
+
+    // check if the product's owner user and the presented authenticated user id are same
+    if (product.user.toString() !== data.user.id) {
+      return NextResponse.json(
+        {
+          success: false,
+          message:
+            "Not allowed to update product, please authenticate with correct credentials",
+        },
         { status: 404 }
       );
     }
@@ -25,7 +45,7 @@ export async function PATCH(request, { params }) {
         {
           success: true,
           message: "Successfull while updating the product with id",
-          product: updatedProduct
+          product: updatedProduct,
         },
         { status: 200 }
       );
@@ -47,6 +67,13 @@ export async function DELETE(request, { params }) {
   try {
     await connectToDB();
     const { id } = params;
+
+    const { data, authenticateResponse } = findIfUserAuthenticated(request);
+    // this will be returned if it is set in the middleware
+    if (authenticateResponse != null) {
+      return authenticateResponse;
+    }
+
     // find if some product with the given id
     const product = await Product.findById(id);
     if (!product) {
@@ -55,6 +82,19 @@ export async function DELETE(request, { params }) {
         { status: 404 }
       );
     }
+
+    // check if the product's owner user and the presented authenticated user id are same
+    if (product.user.toString() !== data.user.id) {
+      return NextResponse.json(
+        {
+          success: false,
+          message:
+            "Not allowed to delete product, please authenticate with correct credentials",
+        },
+        { status: 404 }
+      );
+    }
+
     // as the product exists, so delete the product now
     const deletedProduct = await Product.findByIdAndDelete(id);
     if (deletedProduct) {
